requirements.txt
absl-py==0.3.0
asn1crypto==0.24.0
astor==0.7.1
bleach==2.1.3
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
cryptography==2.3.1
cryptography-vectors==2.3.1
cycler==0.10.0
decorator==4.3.0
entrypoints==0.2.3
eventlet==0.23.0
Flask==1.0.2
gast==0.2.0
greenlet==0.4.13
grpcio==1.14.0
h5py==2.8.0
html5lib==1.0.1
idna==2.7
imageio==2.3.0
ipykernel==4.8.2
ipython==5.0.0
ipython-genutils==0.2.0
ipywidgets==7.4.0
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
mkl-fft==1.0.4
mkl-random==1.0.1
moviepy==0.2.3.5
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.6.0
numpy==1.15.0
olefile==0.45.1
pandas==0.23.4
pandocfilters==1.4.2
pickleshare==0.7.4
Pillow==5.2.0
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.6.0
pycparser==2.18
Pygments==2.2.0
pyOpenSSL==18.0.0
pyparsing==2.2.0
PySocks==1.6.8
python-dateutil==2.7.3
python-engineio==2.2.0
python-socketio==1.9.0
pytz==2018.5
pywinpty==0.5.4
PyYAML==3.13
pyzmq==17.1.0
qtconsole==4.3.1
requests==2.19.1
scikit-learn==0.19.1
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
tensorboard==1.9.0
tensorflow==1.9.0
tensorflow-gpu==1.9.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tornado==5.1
tqdm==4.25.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.0
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
